Better account_names generation + split genesis-create #527
+ generate account names with sha1 and ensure there are no dots
+ move state reading code and constants to separate files
+ rename stake-related structs in abi to conform other names (some, like `param_struct` were too general)
+ improve golos_state accounts converting: there are several paths, how account_names derived from account object/account index/account id. unified helper functions used now instead of random chunks of code
+ enable serializer for system abi too
+ remove obsolete get_type_name helpers
+ remove some debug logs
